requirements.txt
aiohttp>=3.8.0
centrifuge-python>=0.2.0
python-telegram-bot>=20.0
websockets>=11.0